Add --stream option and list of files

+++ cove/management/commands/base_command.py
@@ -2,6 +2,7 @@ import json
 import os
 import shutil
 import sys
+import glob
 
 from django.core.management.base import BaseCommand
 
@@ -19,31 +20,37 @@ class CoveBaseCommand(BaseCommand):
         super(CoveBaseCommand, self).__init__(*args, **kwargs)
 
     def add_arguments(self, parser):
-        parser.add_argument('file', help='File to be processed by Cove')
+        parser.add_argument('files', help='Files to be processed by Cove', nargs='+')
         parser.add_argument('--output-dir', '-o', default='', help='Directory where the output is created, defaults to the name of the file')
         parser.add_argument('--delete', '-d', action='store_true', help='Delete existing directory if it exits')
         parser.add_argument('--exclude-file', '-e', action='store_true', help='Do not include the file in the output directory')
+        parser.add_argument('--stream', '-st', action='store_true', help='Do not include the file in the output directory')
 
-    def handle(self, file, *args, **options):
+    def handle(self, files, *args, **options):
+
+        stream = options.get('stream')
+        if (len(files) > 1 or len(glob.glob(files[0])) > 1 or os.path.isdir(glob.glob(files[0])[0])) and not stream:
+            self.stdout.write('The multiple file option must be used with the --stream option')
+            sys.exit(1)
+
         output_dir = options.get('output_dir')
         delete = options.get('delete')
         exclude_file = options.get('exclude_file')
 
-        if not output_dir:
-            if os.path.isdir(file):
-                output_dir = 'result'
-            else:
+        if not stream:
+            file = files[0]
+            if not output_dir:
                 output_dir = file.split('/')[-1].split('.')[0]
 
-        if os.path.exists(output_dir):
-            if delete:
-                shutil.rmtree(output_dir)
-            else:
-                self.stdout.write('Directory {} already exists'.format(output_dir))
-                sys.exit(1)
-        os.makedirs(output_dir)
+            if os.path.exists(output_dir):
+                if delete:
+                    shutil.rmtree(output_dir)
+                else:
+                    self.stdout.write('Directory {} already exists'.format(output_dir))
+                    sys.exit(1)
+            os.makedirs(output_dir)
 
-        if not exclude_file:
-            shutil.copy2(file, output_dir)
+            if not exclude_file or stream:
+                shutil.copy2(file, output_dir)
 
-        self.output_dir = output_dir
+            self.output_dir = output_dir
